pyproject.toml
[tool.poetry]
name = "ray-processing"
version = "0.1.0"
description = ""
authors = ["mike <xcellentbird@gmail.com>"]
readme = "README.md"
packages = [{include = "ray_processing"}]

[tool.poetry.dependencies]
python = "3.8.13"
ray = {version = "2.8.0", extras = ["default"]}
boto3 = "^1.29.5"
warcio = "^1.7.4"
funcy = "^2.0"
python-dotenv = "^1.0.0"
pyarrow = "^14.0.1"
jupyter = "^1.0.0"
pandas = ">=1.4.3"
pillow = "^10.1.0"
tqdm = "^4.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
